move SendReceive tests to separate files
