+++ src/libraries/System.Net.Sockets/tests/FunctionalTests/SendReceive/SendReceiveListener.cs
@@ -1,0 +1,91 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.DotNet.RemoteExecutor;
+using Microsoft.DotNet.XUnitExtensions;
+using Xunit;
+using Xunit.Abstractions;
+using Xunit.Sdk;
+
+
+namespace System.Net.Sockets.Tests
+{
+    public sealed class SendReceiveListener : MemberDatas
+    {
+        [OuterLoop]
+        [Theory]
+        [MemberData(nameof(Loopbacks))]
+        public async Task SendRecvAsync_TcpListener_TcpClient(IPAddress listenAt)
+        {
+            const int BytesToSend = 123456;
+            const int ListenBacklog = 1;
+            const int LingerTime = 10;
+            const int TestTimeout = 30000;
+
+            var listener = new TcpListener(listenAt, 0);
+            listener.Start(ListenBacklog);
+
+            int bytesReceived = 0;
+            var receivedChecksum = new Fletcher32();
+            Task serverTask = Task.Run(async () =>
+            {
+                using (TcpClient remote = await listener.AcceptTcpClientAsync())
+                using (NetworkStream stream = remote.GetStream())
+                {
+                    var recvBuffer = new byte[256];
+                    while (true)
+                    {
+                        int received = await stream.ReadAsync(recvBuffer, 0, recvBuffer.Length);
+                        if (received == 0)
+                        {
+                            break;
+                        }
+
+                        bytesReceived += received;
+                        receivedChecksum.Add(recvBuffer, 0, received);
+                    }
+                }
+            });
+
+            int bytesSent = 0;
+            var sentChecksum = new Fletcher32();
+            Task clientTask = Task.Run(async () =>
+            {
+                var clientEndpoint = (IPEndPoint)listener.LocalEndpoint;
+
+                using (var client = new TcpClient(clientEndpoint.AddressFamily))
+                {
+                    await client.ConnectAsync(clientEndpoint.Address, clientEndpoint.Port);
+
+                    using (NetworkStream stream = client.GetStream())
+                    {
+                        var random = new Random();
+                        var sendBuffer = new byte[512];
+                        for (int remaining = BytesToSend, sent = 0; remaining > 0; remaining -= sent)
+                        {
+                            random.NextBytes(sendBuffer);
+
+                            sent = Math.Min(sendBuffer.Length, remaining);
+                            await stream.WriteAsync(sendBuffer, 0, sent);
+
+                            bytesSent += sent;
+                            sentChecksum.Add(sendBuffer, 0, sent);
+                        }
+
+                        client.LingerState = new LingerOption(true, LingerTime);
+                    }
+                }
+            });
+
+            await (new[] { serverTask, clientTask }).WhenAllOrAnyFailed(TestTimeout);
+
+            Assert.Equal(bytesSent, bytesReceived);
+            Assert.Equal(sentChecksum.Sum, receivedChecksum.Sum);
+        }
+    }
+}
